pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.12",
    "anthropic>=0.45.2",
    "beautifulsoup4>=4.13.3",
    "cloudscraper>=1.2.71",
    "fake-useragent>=2.0.3",
    "flask>=3.1.0",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "openai>=1.63.0",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.6",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.1",
    "requests[socks]>=2.32.3",
    "selenium>=4.28.1",
    "sqlalchemy>=2.0.38",
    "streamlit>=1.42.0",
    "tenacity>=9.0.0",
    "trafilatura>=2.0.0",
    "twilio>=9.4.5",
    "undetected-chromedriver>=3.5.5",
    "urllib3>=2.3.0",
    "webdriver-manager>=4.0.2",
    "xlsxwriter>=3.2.2",
]
